+++ riptide_controllers/src/riptide_controllers/controller_overseer.py
@@ -3,6 +3,7 @@
 import os
 import yaml
 import yaml.parser
+from importlib import import_module
 import rclpy
 import numpy as np
 from collections import deque
@@ -23,35 +24,8 @@ from std_srvs.srv import Trigger, SetBool
 from diagnostic_msgs.msg import DiagnosticArray
 from geometry_msgs.msg import Twist, TwistWithCovariance
 
-
-#TODO: sort these
-ERROR_PATIENCE = 1.0
-
-#the number of frames the escs can be powered off before clearing acculators
-ESC_POWER_STOP_TOLERANCE = 2
-ESC_POWER_TIMEOUT = 2
-
-PARAMETER_SCALE = 1000000
-
-THRUSTER_SOLVER_WEIGHT_MATRIX_TOPIC = "controller/solver_weights"
-
-#paramters that cannot be pulled from yaml but still need setting
-SPECIAL_PARAMETERS = ["talos_wrenchmat"]
-
-ACTIVE_PARAMETERS_MASK = "controller__active_force_mask"
-
-FF_PUBLISH_PARAM = "disable_native_ff"
-FF_TOPIC_NAME = "controller/FF_body_force"
-AUTOTUNE_REINIT_TOPIC_NAME = "controller/re_init_accumulators"
-WEIGHTS_FORCE_UPDATE_PERIOD = 1
-
-ORIN_AUTOTUNE_DIR = "/bin"
-
-AUTOFF_INIT_TOLERANCE = .01 #seeing rounding errors lol
-
-G = 9.8067
-
-RELOAD_TIME = 2 # seconds
+from overseer_global_defs import \
+    PARAMETER_SCALE, SPECIAL_PARAMETERS, RELOAD_TIME, ORIN_AUTOTUNE_DIR, ACTIVE_PARAMETERS_MASK
 
 
 # wrapper class for rclpy service which monitors for timeouts and prevents concurrent operations
@@ -111,18 +85,25 @@ class MonitoredServiceClient():
 
 #manages an individual simulink model.
 class SimulinkModelNode():
-    def __init__(self, node: 'ControllerOverseer', node_name: str):
+    def __init__(self, node: 'ControllerOverseer', node_name: str, config, plugin_names: list[str]):
         self.model_active = False
         self.full_node_name = ""
         self.list_param_client = None
         self.set_param_client = None
         self.overseer_node = node
         self.node_name = node_name
+        self.config = config
         self.known_params = []
         self.reload_param_service = node.create_service(Trigger, f"controller_overseer/update_{self.node_name}_params".lower(), self.reload_parameters_callback)
         self.last_reload_time = node.get_clock().now()
         self.loaded_params = False # have parameters been loaded
-
+        self.loaded_plugins = False # have plugins been loaded
+        
+        # initialize plugins, but dont load them yet
+        self.plugins = {}
+        for plugin_name in plugin_names:
+            factory = getattr(import_module(plugin_name), "get_plugin")
+            self.plugins[plugin_name] = factory()
     
     # checks if the model is active. Handles when model comes up or goes down
     # IMPORTANT: active_nodes must be a list of the FULL node names (e.g. /talos/...).
@@ -141,6 +122,11 @@ class SimulinkModelNode():
                 #note the model is active
                 self.model_active = True                
                 self.overseer_node.get_logger().info(f"Found {self.node_name} as {self.full_node_name}!")
+                
+                #model is present. now init required plugins if not already loaded
+                if not self.loaded_plugins:
+                    for plugin in self.plugins:
+                        self.plugins[plugin].init_plugin(self.overseer_node, self.config, self)
 
                 #the client to set the params of the model
                 self.list_param_client = MonitoredServiceClient(self.overseer_node, ListParameters, f"{self.full_node_name}/list_parameters")
@@ -253,13 +239,7 @@ class SimulinkModelNode():
         #check to see if the param is special
         if param_name in SPECIAL_PARAMETERS:
             if param_name == "talos_wrenchmat":
-                #change to 1D array
-                effects_1D = []
-                for column in self.overseer_node.thrusterEffects:
-                    for effect in column:
-                        effects_1D.append(effect)
-
-                return effects_1D, list
+                return self.plugins["overseer_thruster_solver_plugin"].get1dWrenchmat()
             else:
                 self.overseer_node.get_logger().warn("Undefine special parameter!")
 
@@ -311,21 +291,21 @@ class SimulinkModelNode():
 
 #manages parameters for the controller / thruster solver
 class ControllerOverseer(Node):
-    waiting_on_init = False
+    # waiting_on_init = False
 
-    drag_comp_forward_data = None
-    drag_comp_reverse_data = None
+    # drag_comp_forward_data = None
+    # drag_comp_reverse_data = None
 
-    refresh_drag_file = True
+    # refresh_drag_file = True
 
     def __init__(self):
-        super().__init__("controllerOverseer")
+        super().__init__("controller_overseer")
         
-        self.escPowerStopsLow = 0
-        self.escPowerStopsHigh = 0
-        self.configTree = {}
-        self.currentAutoTuneTwist = [0,0,0,0,0,0]
-        self.autoff_config_path = None
+        # self.escPowerStopsLow = 0
+        # self.escPowerStopsHigh = 0
+        # self.configTree = {}
+        # self.currentAutoTuneTwist = [0,0,0,0,0,0]
+        # self.autoff_config_path = None
 
         #
         # Declare parameters 
@@ -339,15 +319,15 @@ class ControllerOverseer(Node):
         self.declare_parameter("vehicle_config", "")
         self.config_path = self.get_parameter("vehicle_config").value
 
-        #get thruster solver node name
-        self.declare_parameter("thruster_solver_node_name", "")
-        self.thrusterSolverName = self.get_parameter("thruster_solver_node_name").value
+        # #get thruster solver node name
+        # self.declare_parameter("thruster_solver_node_name", "")
+        # self.thrusterSolverName = self.get_parameter("thruster_solver_node_name").value
 
         #declare write autotune param
-        self.declare_parameter("write_ff_autotune", True)
-        self.writeFFAutoTune = self.get_parameter("write_ff_autotune").value
+        # self.declare_parameter("write_ff_autotune", True)
+        # self.writeFFAutoTune = self.get_parameter("write_ff_autotune").value
 
-        self.declare_parameter(FF_PUBLISH_PARAM, False)
+        # self.declare_parameter(FF_PUBLISH_PARAM, False)
         
         # Other setup
         
@@ -359,113 +339,116 @@ class ControllerOverseer(Node):
         
         # tracked simulink models
         self.model_nodes = {
-            "thruster_solver": SimulinkModelNode(self, "thruster_solver"),
-            "PID": SimulinkModelNode(self, "PID"),
-            "SMC": SimulinkModelNode(self, "SMC"),
-            "complete_controller": SimulinkModelNode(self, "complete_controller")
+            "complete_controller": SimulinkModelNode(self, "complete_controller", self.configTree,
+                [
+                    "overseer_6dof_teleop_plugin",
+                    "overseer_autotune_plugin",
+                    "overseer_ff_plugin",
+                    "overseer_thruster_solver_plugin"
+                ])
         }
 
         #generate thruster force matrix
-        self.generateThrusterForceMatrix(self.thruster_info, self.com)
+        # self.generateThrusterForceMatrix(self.thruster_info, self.com)
 
         #default thruster weights and working thrusters
-        self.activeThrusters = [True, True, True, True, True, True, True, True]
-        self.submergedThrusters = [True, True, True, True, True, True, True, True]
-        self.thrusterWeights = [1, 1, 1, 1, 1, 1, 1, 1]
+        # self.activeThrusters = [True, True, True, True, True, True, True, True]
+        # self.submergedThrusters = [True, True, True, True, True, True, True, True]
+        # self.thrusterWeights = [1, 1, 1, 1, 1, 1, 1, 1]
 
         # the thruster mode
-        self.thrusterMode = 1
+        # self.thrusterMode = 1
 
         #declare pubs and subs
 
         #thruster telemetry
-        self.create_subscription(DshotPartialTelemetry, "state/thrusters/telemetry", self.thrusterTelemetryCB, qos_profile_system_default)
+        # self.create_subscription(DshotPartialTelemetry, "state/thrusters/telemetry", self.thrusterTelemetryCB, qos_profile_system_default)
 
         #publish to active controllers whether motion is enabled
-        self.motionEnabledPub = self.create_publisher(Bool, "controller/motion_enabled", qos_profile=qos_profile_system_default)
+        # self.motionEnabledPub = self.create_publisher(Bool, "controller/motion_enabled", qos_profile=qos_profile_system_default)
 
         #thruster solver parameters
-        self.setThrusterSolverParamsClient = self.create_client(SetParameters, f"{self.thrusterSolverName}/set_parameters")
+        # self.setThrusterSolverParamsClient = self.create_client(SetParameters, f"{self.thrusterSolverName}/set_parameters")
 
         #thruster mode
-        self.create_subscription(Int16, "thrusterSolver/thrusterState", self.setThrusterModeCB, qos_profile_system_default)
+        # self.create_subscription(Int16, "thrusterSolver/thrusterState", self.setThrusterModeCB, qos_profile_system_default)
 
         #odometry filtered
-        self.create_subscription(Odometry, "odometry/filtered", self.odometryCB, qos_profile_system_default)
+        # self.create_subscription(Odometry, "odometry/filtered", self.odometryCB, qos_profile_system_default)
 
         #sub to autotune
-        self.create_subscription(Twist, "ff_auto_tune", self.ffAutoTuneCB, qos_profile_system_default)
+        # self.create_subscription(Twist, "ff_auto_tune", self.ffAutoTuneCB, qos_profile_system_default)
 
-        #sub to the drag compensator forward
-        self.create_subscription(TwistWithCovariance, "controller/drag_comp/forward", self.dragCompForwardCb, qos_profile_system_default)
+        # #sub to the drag compensator forward
+        # self.create_subscription(TwistWithCovariance, "controller/drag_comp/forward", self.dragCompForwardCb, qos_profile_system_default)
 
-        #sub to the drag compensator reverse
-        self.create_subscription(TwistWithCovariance, "controller/drag_comp/reverse", self.dragCompRevereseCb, qos_profile_system_default)
+        # #sub to the drag compensator reverse
+        # self.create_subscription(TwistWithCovariance, "controller/drag_comp/reverse", self.dragCompRevereseCb, qos_profile_system_default)
 
         #pub for thruster weights
-        self.weightsPub = self.create_publisher(Int32MultiArray, THRUSTER_SOLVER_WEIGHT_MATRIX_TOPIC, qos_profile_system_default)
+        # self.weightsPub = self.create_publisher(Int32MultiArray, THRUSTER_SOLVER_WEIGHT_MATRIX_TOPIC, qos_profile_system_default)
 
         #pub ff force
-        self.ffPub = self.create_publisher(Twist, FF_TOPIC_NAME, qos_profile_system_default)
+        # self.ffPub = self.create_publisher(Twist, FF_TOPIC_NAME, qos_profile_system_default)
 
         #pub signal to reinit autoff and dragcal
-        self.re_init_signal_pub = self.create_publisher(Empty, AUTOTUNE_REINIT_TOPIC_NAME, qos_profile_system_default)
+        # self.re_init_signal_pub = self.create_publisher(Empty, AUTOTUNE_REINIT_TOPIC_NAME, qos_profile_system_default)
 
         #create teleop serivce
-        self.setTeleop = self.create_service(SetBool, "setTeleop", self.setTeleop)
+        # self.setTeleop = self.create_service(SetBool, "setTeleop", self.setTeleop)
 
         #declare transform
-        self.tfBuffer = Buffer()
-        self.tfListener = TransformListener(self.tfBuffer, self)
+        # self.tfBuffer = Buffer()
+        # self.tfListener = TransformListener(self.tfBuffer, self)
 
-        #the tf namespace
-        # TODO: FIX
-        # self.tfNamespace = self.get_parameter("robot").value
-        self.tfNamespace = "talos"
+        # #the tf namespace
+        # # TODO: FIX
+        # # self.tfNamespace = self.get_parameter("robot").value
+        # self.tfNamespace = "talos"
 
         #start time
-        self.startTime = None
+        # self.startTime = None
 
         #timer to update models and ff
         self.create_timer(1.0, self.doUpdate)
 
         #start the publish loop
-        self.create_timer(WEIGHTS_FORCE_UPDATE_PERIOD, self.adjustThrusterWeights)
+        # self.create_timer(WEIGHTS_FORCE_UPDATE_PERIOD, self.adjustThrusterWeights)
 
-        self.enabled = True
+        # self.enabled = True
         self.publishingFF = True
 
-        #a timer to ensure that the active controllers stop if telemetry stops publishing!
-        self.escPowerCheckTimer = self.create_timer(ESC_POWER_TIMEOUT, self.escPowerTimeout)
+        # # a timer to ensure that the active controllers stop if telemetry stops publishing!
+        # self.escPowerCheckTimer = self.create_timer(ESC_POWER_TIMEOUT, self.escPowerTimeout)
 
 
-    def dragCompForwardCb(self, msg):
+    # def dragCompForwardCb(self, msg):
 
-        #save drag data
-        self.drag_comp_forward_data = msg.covariance
-        self.refresh_drag_file = True
+    #     #save drag data
+    #     self.drag_comp_forward_data = msg.covariance
+    #     self.refresh_drag_file = True
 
-    def dragCompRevereseCb(self, msg):
+    # def dragCompRevereseCb(self, msg):
 
-        #save drag data
-        self.drag_comp_reverse_data = msg.covariance
-        self.refresh_drag_file = True
+    #     #save drag data
+    #     self.drag_comp_reverse_data = msg.covariance
+    #     self.refresh_drag_file = True
 
 
 
-    def generateThrusterForceMatrix(self, thruster_info, com):
-        #generate the thruster effect matrix
-        self.thrusterEffects = np.zeros(shape=(8,6))
-        for i, thruster in enumerate(thruster_info):
-            pose = np.array(thruster["pose"])
+    # def generateThrusterForceMatrix(self, thruster_info, com):
+    #     #generate the thruster effect matrix
+    #     self.thrusterEffects = np.zeros(shape=(8,6))
+    #     for i, thruster in enumerate(thruster_info):
+    #         pose = np.array(thruster["pose"])
 
-            #calculate the force vector generated by the thruster
-            forceVector = np.matmul(euler_matrix(pose[3], pose[4], pose[5])[:3, :3], np.array([1,0,0]))
-            positionFromCOM = pose[:3] - com
-            torque = np.cross(positionFromCOM[:3], forceVector)
+    #         #calculate the force vector generated by the thruster
+    #         forceVector = np.matmul(euler_matrix(pose[3], pose[4], pose[5])[:3, :3], np.array([1,0,0]))
+    #         positionFromCOM = pose[:3] - com
+    #         torque = np.cross(positionFromCOM[:3], forceVector)
 
-            #insert into thruster effect matrix
-            self.thrusterEffects[i] = [forceVector[0], forceVector[1], forceVector[2], torque[0], torque[1], torque[2]]
+    #         #insert into thruster effect matrix
+    #         self.thrusterEffects[i] = [forceVector[0], forceVector[1], forceVector[2], torque[0], torque[1], torque[2]]
 
     def setConfigPath(self):
         #set the path to the config file
@@ -552,279 +535,276 @@ class ControllerOverseer(Node):
             self.get_logger().error(f"Type Error!")
         except yaml.parser.ParserError:
             self.get_logger().error(f"Yaml Parser")
-
-
-            
         
         # read specific values used by the class
                 
         # thruster info
-        self.thruster_info = self.configTree['thrusters']
+        # self.thruster_info = self.configTree['thrusters']
         
         #read in kill plane height
-        self.killPlane = self.configTree["controller_overseer"]["thruster_kill_plane"]
+        # self.killPlane = self.configTree["controller_overseer"]["thruster_kill_plane"]
         
         # read in com data
-        self.com = self.configTree["com"]
+        # self.com = self.configTree["com"]
         
         # feed-forward
-        self.base_wrench = self.configTree["controller"]["feed_forward"]["base_wrench"]
+        # self.base_wrench = self.configTree["controller"]["feed_forward"]["base_wrench"]
         
         #thruster solver
-        thruster_solver_info = self.configTree["thruster_solver"]
+        # thruster_solver_info = self.configTree["thruster_solver"]
         
-        #read in weight info
-        self.defaultWeight      = thruster_solver_info["default_weight"]
-        self.surfaceWeight      = thruster_solver_info["surfaced_weight"]
-        self.disabledWeight     = thruster_solver_info["disable_weight"]
-        self.lowDowndraftWeight = thruster_solver_info["low_downdraft_weight"]
+        # #read in weight info
+        # self.defaultWeight      = thruster_solver_info["default_weight"]
+        # self.surfaceWeight      = thruster_solver_info["surfaced_weight"]
+        # self.disabledWeight     = thruster_solver_info["disable_weight"]
+        # self.lowDowndraftWeight = thruster_solver_info["low_downdraft_weight"]
 
 
-    def thrusterTelemetryCB(self, msg: DshotPartialTelemetry):
-        #wether or not the thrusterSolverweigths need adjusted
-        adjustWeights = False
+    # def thrusterTelemetryCB(self, msg: DshotPartialTelemetry):
+    #     #wether or not the thrusterSolverweigths need adjusted
+    #     adjustWeights = False
 
-        #restart the timeout
-        self.escPowerCheckTimer.reset()
+    #     #restart the timeout
+    #     self.escPowerCheckTimer.reset()
 
-        # which thrusters - groups of 4
-        if msg.start_thruster_num == 0:
-            #check wether each thruster is active
-            for i, esc in enumerate(msg.esc_telemetry):
-                if not esc.thruster_ready:
-                    if self.activeThrusters[i] == True:
-                        #if change, adjust the thruster weights
-                        self.activeThrusters[i] = False
-                        adjustWeights = True
-                else:
-                    if self.activeThrusters[i] == False:
-                        self.activeThrusters[i] = True
-                        adjustWeights = True
+    #     # which thrusters - groups of 4
+    #     if msg.start_thruster_num == 0:
+    #         #check wether each thruster is active
+    #         for i, esc in enumerate(msg.esc_telemetry):
+    #             if not esc.thruster_ready:
+    #                 if self.activeThrusters[i] == True:
+    #                     #if change, adjust the thruster weights
+    #                     self.activeThrusters[i] = False
+    #                     adjustWeights = True
+    #             else:
+    #                 if self.activeThrusters[i] == False:
+    #                     self.activeThrusters[i] = True
+    #                     adjustWeights = True
 
-            #check if the boards are enebaled
-            if not (msg.disabled_flags == 0):  
-                self.escPowerStopsLow += 1    
-            else:
-                self.escPowerStopsLow = 0 
+    #         #check if the boards are enebaled
+    #         if not (msg.disabled_flags == 0):  
+    #             self.escPowerStopsLow += 1    
+    #         else:
+    #             self.escPowerStopsLow = 0 
 
-        else:
-            for i, esc in enumerate(msg.esc_telemetry):
-                if not esc.thruster_ready:
-                    if self.activeThrusters[i + 4] == True:
-                        self.activeThrusters[i + 4] = False
-                        adjustWeights = True
-                else:
-                    if self.activeThrusters[i + 4] == False:
-                        self.activeThrusters[i + 4] = True
-                        adjustWeights = True
+    #     else:
+    #         for i, esc in enumerate(msg.esc_telemetry):
+    #             if not esc.thruster_ready:
+    #                 if self.activeThrusters[i + 4] == True:
+    #                     self.activeThrusters[i + 4] = False
+    #                     adjustWeights = True
+    #             else:
+    #                 if self.activeThrusters[i + 4] == False:
+    #                     self.activeThrusters[i + 4] = True
+    #                     adjustWeights = True
 
-            #check if the boards are enebaled
-            if not (msg.disabled_flags == 0):  
-                self.escPowerStopsHigh += 1  
-            else:
-                self.escPowerStopsHigh = 0     
+    #         #check if the boards are enebaled
+    #         if not (msg.disabled_flags == 0):  
+    #             self.escPowerStopsHigh += 1  
+    #         else:
+    #             self.escPowerStopsHigh = 0     
 
-        #adjust the weights of the thruster solver if anything has changed
-        if adjustWeights:
-            self.adjustThrusterWeights()
+    #     #adjust the weights of the thruster solver if anything has changed
+    #     if adjustWeights:
+    #         self.adjustThrusterWeights()
 
-        motionMsg = Bool()
-        #check if the boards are enebaled
-        if self.escPowerStopsLow > ESC_POWER_STOP_TOLERANCE or self.escPowerStopsHigh > ESC_POWER_STOP_TOLERANCE:
-            #publish disabled message
-            motionMsg.data = False
-            self.enabled = False
+    #     motionMsg = Bool()
+    #     #check if the boards are enebaled
+    #     if self.escPowerStopsLow > ESC_POWER_STOP_TOLERANCE or self.escPowerStopsHigh > ESC_POWER_STOP_TOLERANCE:
+    #         #publish disabled message
+    #         motionMsg.data = False
+    #         self.enabled = False
 
-            # if self.enabled:
-            #     self.get_logger().warn("Recieving disabled flags from ESC!")
-        else:
+    #         # if self.enabled:
+    #         #     self.get_logger().warn("Recieving disabled flags from ESC!")
+    #     else:
 
-            motionMsg.data = True
-            self.enabled = True
-            #publish enabled message
+    #         motionMsg.data = True
+    #         self.enabled = True
+    #         #publish enabled message
         
-        self.motionEnabledPub.publish(motionMsg)
+    #     self.motionEnabledPub.publish(motionMsg)
 
 
-    def escPowerTimeout(self):
-        #timeout for if the escs go to long without publishing telemerty
-        if self.enabled:
-            self.get_logger().warn("Not recieving thruster telemetry!")
+    # def escPowerTimeout(self):
+    #     #timeout for if the escs go to long without publishing telemerty
+    #     if self.enabled:
+    #         self.get_logger().warn("Not recieving thruster telemetry!")
 
-        motionMsg = Bool()
-        motionMsg.data = False
-        self.enabled = False
-        self.motionEnabledPub.publish(motionMsg)
-
-
-    def setThrusterModeCB(self, msg:Int16):
-        #change the thruster mode
-
-        #Modes -----
-        #   1 - Normal
-        #   2 - Low Downdraft
-
-        if not (msg.data == self.thrusterMode):
-            self.thrusterMode = msg.data
-
-            #update the weights
-            self.adjustThrusterWeights()
+    #     motionMsg = Bool()
+    #     motionMsg.data = False
+    #     self.enabled = False
+    #     self.motionEnabledPub.publish(motionMsg)
 
 
-    def odometryCB(self, msg):
-        #check if thrusters are submerged - everytime odom is updated - just using as a frequency
+    # def setThrusterModeCB(self, msg:Int16):
+    #     #change the thruster mode
 
-        #start the start time on first cb
-        if self.startTime is None:
-            self.startTime = self.get_clock().now()
+    #     #Modes -----
+    #     #   1 - Normal
+    #     #   2 - Low Downdraft
 
-        submerged = [False, False, False, False, False, False, False, False]
+    #     if not (msg.data == self.thrusterMode):
+    #         self.thrusterMode = msg.data
 
-        try:
-            #look at each thrusterbreak
-            for i, thursterSufaced in enumerate(self.submergedThrusters):
-                pos = self.tfBuffer.lookup_transform("world", f"{self.tfNamespace}/thruster_{i}", Time())
-
-                #if thruster is above the kill plane
-                if pos.transform.translation.z < self.killPlane:
-                    submerged[i] = True
-
-            if not (np.array_equal(submerged, self.submergedThrusters)):
-                #if a different thruster combo is submerdged, adjust, the weights
-                self.submergedThrusters = submerged
-                self.adjustThrusterWeights()
-
-        except Exception as ex:
-            if self.get_clock().now().to_msg().sec >= ERROR_PATIENCE + self.startTime.to_msg().sec:
-                self.get_logger().error("Thruster Position Lookup failed with exception: " + str(ex))
+    #         #update the weights
+    #         self.adjustThrusterWeights()
 
 
-    def adjustThrusterWeights(self):
-        #TODO add weight values into descriptions
+    # def odometryCB(self, msg):
+    #     #check if thrusters are submerged - everytime odom is updated - just using as a frequency
 
-        #number of active thrusters
-        activeThrusterCount = 0
-        submergedThrusters = 0
+    #     #start the start time on first cb
+    #     if self.startTime is None:
+    #         self.startTime = self.get_clock().now()
 
-        #shutoff inactive thrusters - disabled or broken
-        for i, isActive in enumerate(self.activeThrusters):
-            if isActive:
-                activeThrusterCount += 1
+    #     submerged = [False, False, False, False, False, False, False, False]
 
-                #play around with weights for thrusters above surface
-                if not (self.submergedThrusters[i] == True):
-                    #if thruster is not submerdged
-                    self.thrusterWeights[i] = self.surfaceWeight
-                else:
-                    #thruster is active and submerdged
-                    submergedThrusters += 1
-                    self.thrusterWeights[i] = self.defaultWeight
+    #     try:
+    #         #look at each thrusterbreak
+    #         for i, thursterSufaced in enumerate(self.submergedThrusters):
+    #             pos = self.tfBuffer.lookup_transform("world", f"{self.tfNamespace}/thruster_{i}", Time())
 
-            else:
-                #if a thruster is inactive - raise the cost of "using" thruster
-                self.thrusterWeights[i] = self.disabledWeight
+    #             #if thruster is above the kill plane
+    #             if pos.transform.translation.z < self.killPlane:
+    #                 submerged[i] = True
 
-        if activeThrusterCount <= 6:
-            #if system is not full actuated, it cannot be optimized, very high rpms / force can be requested
-            #for the safety of the system, this will autodisable robot (probably)
-            #remove if PIA
-            if self.enabled:
-                self.get_logger().error("System has become underactuated. Only:  " + str(activeThrusterCount) + " thrusters are active. Killing Thrusters!")
-                self.enabled = False
-        else:
-            self.enabled = True
+    #         if not (np.array_equal(submerged, self.submergedThrusters)):
+    #             #if a different thruster combo is submerdged, adjust, the weights
+    #             self.submergedThrusters = submerged
+    #             self.adjustThrusterWeights()
 
-        if submergedThrusters >= 8:
-            #take into account control modes only if all actuators are working
+    #     except Exception as ex:
+    #         if self.get_clock().now().to_msg().sec >= ERROR_PATIENCE + self.startTime.to_msg().sec:
+    #             self.get_logger().error("Thruster Position Lookup failed with exception: " + str(ex))
 
-            if self.thrusterMode == 2:
-                #apply low downdraft
-                self.thrusterWeights[4] = self.lowDowndraftWeight
-                self.thrusterWeights[5] = self.lowDowndraftWeight
 
-        #impose disable weight if nessecary
-        if self.thrusterMode == 0:
-            self.thrusterWeights = [0,0,0,0,0,0,0,0]
+    # def adjustThrusterWeights(self):
+    #     #TODO add weight values into descriptions
 
-        msg = Int32MultiArray()
+    #     #number of active thrusters
+    #     activeThrusterCount = 0
+    #     submergedThrusters = 0
 
-        #scale and round all weights
-        weights = []
-        for weight in self.thrusterWeights:
-            weights.append(int(weight))
+    #     #shutoff inactive thrusters - disabled or broken
+    #     for i, isActive in enumerate(self.activeThrusters):
+    #         if isActive:
+    #             activeThrusterCount += 1
 
-        msg.data = weights
+    #             #play around with weights for thrusters above surface
+    #             if not (self.submergedThrusters[i] == True):
+    #                 #if thruster is not submerdged
+    #                 self.thrusterWeights[i] = self.surfaceWeight
+    #             else:
+    #                 #thruster is active and submerdged
+    #                 submergedThrusters += 1
+    #                 self.thrusterWeights[i] = self.defaultWeight
 
-        #publish weights
-        self.weightsPub.publish(msg)
+    #         else:
+    #             #if a thruster is inactive - raise the cost of "using" thruster
+    #             self.thrusterWeights[i] = self.disabledWeight
 
-    def setTeleop(self, request, future):
-        #set the teleop mode
+    #     if activeThrusterCount <= 6:
+    #         #if system is not full actuated, it cannot be optimized, very high rpms / force can be requested
+    #         #for the safety of the system, this will autodisable robot (probably)
+    #         #remove if PIA
+    #         if self.enabled:
+    #             self.get_logger().error("System has become underactuated. Only:  " + str(activeThrusterCount) + " thrusters are active. Killing Thrusters!")
+    #             self.enabled = False
+    #     else:
+    #         self.enabled = True
 
-        #if setting Teleop on
-        if(request.data == True):
-            try:
-                #set the control mask values
-                pVal = ParameterValue()
-                pVal.type = ParameterType.PARAMETER_INTEGER_ARRAY
-                pVal.integer_array_value = [3000000, 3000000, 2000000, 1000000, 1000000, 3000000]
+    #     if submergedThrusters >= 8:
+    #         #take into account control modes only if all actuators are working
 
-                param = Parameter()
-                param.value = pVal
-                param.name = ACTIVE_PARAMETERS_MASK
+    #         if self.thrusterMode == 2:
+    #             #apply low downdraft
+    #             self.thrusterWeights[4] = self.lowDowndraftWeight
+    #             self.thrusterWeights[5] = self.lowDowndraftWeight
 
-                request = SetParameters.Request()
-                request.parameters = [param]
+    #     #impose disable weight if nessecary
+    #     if self.thrusterMode == 0:
+    #         self.thrusterWeights = [0,0,0,0,0,0,0,0]
 
-                self.model_nodes["complete_controller"].set_param_client.schedule_call(request, self.model_nodes["complete_controller"].set_parameters_done_callback)
+    #     msg = Int32MultiArray()
 
-                future.success = True
-                future.message = "Successfully enabled teleop!"
+    #     #scale and round all weights
+    #     weights = []
+    #     for weight in self.thrusterWeights:
+    #         weights.append(int(weight))
 
-            except Exception as E:
+    #     msg.data = weights
 
-                if(type(E) == AttributeError):
-                    #this is an expected error so elaborate
-                    self.get_logger().warn("Failed to initialize teleop, model is not started!")
-                    future.success = False
-                    future.message = "Failed to initialize teleop, model is not started!"
-                else:
-                    self.get_logger().warn(f"Failed to initialize teleop! {E}")
-                    future.success = False
-                    future.message = f"Failed to initialize teleop! {E}" 
-        else:
-        #putting into active control
-            try:
-                #set the control mask values
-                pVal = ParameterValue()
-                pVal.type = ParameterType.PARAMETER_INTEGER_ARRAY
-                pVal.integer_array_value = [2000000, 2000000, 2000000, 1000000, 1000000, 2000000]
+    #     #publish weights
+    #     self.weightsPub.publish(msg)
 
-                param = Parameter()
-                param.value = pVal
-                param.name = ACTIVE_PARAMETERS_MASK
+    # def setTeleop(self, request, future):
+    #     #set the teleop mode
 
-                request = SetParameters.Request()
-                request.parameters = [param]
+    #     #if setting Teleop on
+    #     if(request.data == True):
+    #         try:
+    #             #set the control mask values
+    #             pVal = ParameterValue()
+    #             pVal.type = ParameterType.PARAMETER_INTEGER_ARRAY
+    #             pVal.integer_array_value = [3000000, 3000000, 2000000, 1000000, 1000000, 3000000]
 
-                self.model_nodes["complete_controller"].set_param_client.schedule_call(request, self.model_nodes["complete_controller"].set_parameters_done_callback)
+    #             param = Parameter()
+    #             param.value = pVal
+    #             param.name = ACTIVE_PARAMETERS_MASK
 
-                future.success = True
-                future.message = "Successfully enabled active control!"
+    #             request = SetParameters.Request()
+    #             request.parameters = [param]
 
-            except Exception as E:
+    #             self.model_nodes["complete_controller"].set_param_client.schedule_call(request, self.model_nodes["complete_controller"].set_parameters_done_callback)
 
-                if(type(E) == AttributeError):
-                    #this is an expected error so elaborate
-                    self.get_logger().warn("Failed to initialize active control, model is not started!")
-                    future.success = False
-                    future.message = "Failed to initialize active control, model is not started!"
-                else:
-                    self.get_logger().warn(f"Failed to initialize active control! {E}")
-                    future.success = False
-                    future.message = f"Failed to initialize active control! {E}" 
+    #             future.success = True
+    #             future.message = "Successfully enabled teleop!"
+
+    #         except Exception as E:
+
+    #             if(type(E) == AttributeError):
+    #                 #this is an expected error so elaborate
+    #                 self.get_logger().warn("Failed to initialize teleop, model is not started!")
+    #                 future.success = False
+    #                 future.message = "Failed to initialize teleop, model is not started!"
+    #             else:
+    #                 self.get_logger().warn(f"Failed to initialize teleop! {E}")
+    #                 future.success = False
+    #                 future.message = f"Failed to initialize teleop! {E}" 
+    #     else:
+    #     #putting into active control
+    #         try:
+    #             #set the control mask values
+    #             pVal = ParameterValue()
+    #             pVal.type = ParameterType.PARAMETER_INTEGER_ARRAY
+    #             pVal.integer_array_value = [2000000, 2000000, 2000000, 1000000, 1000000, 2000000]
+
+    #             param = Parameter()
+    #             param.value = pVal
+    #             param.name = ACTIVE_PARAMETERS_MASK
+
+    #             request = SetParameters.Request()
+    #             request.parameters = [param]
+
+    #             self.model_nodes["complete_controller"].set_param_client.schedule_call(request, self.model_nodes["complete_controller"].set_parameters_done_callback)
+
+    #             future.success = True
+    #             future.message = "Successfully enabled active control!"
+
+    #         except Exception as E:
+
+    #             if(type(E) == AttributeError):
+    #                 #this is an expected error so elaborate
+    #                 self.get_logger().warn("Failed to initialize active control, model is not started!")
+    #                 future.success = False
+    #                 future.message = "Failed to initialize active control, model is not started!"
+    #             else:
+    #                 self.get_logger().warn(f"Failed to initialize active control! {E}")
+    #                 future.success = False
+    #                 future.message = f"Failed to initialize active control! {E}" 
                 
-        return future
+    #     return future
             
 
     def doUpdate(self):
@@ -838,111 +818,112 @@ class ControllerOverseer(Node):
         for model_node in self.model_nodes.values():
             model_node.check_if_active(active_rosnode_names)        
         
-        # publish ff if necessary
-        if not (self.get_parameter(FF_PUBLISH_PARAM).value):
-            #publish the ff
-            self.publishingFF = True
-
-            msg = Twist()
-            msg.linear.x = self.base_wrench[0]
-            msg.linear.y = self.base_wrench[1]
-            msg.linear.z = self.base_wrench[2]
-            msg.angular.x = self.base_wrench[3]
-            msg.angular.y = self.base_wrench[4]
-            msg.angular.z = self.base_wrench[5]
-
-            self.ffPub.publish(msg)
-        elif self.publishingFF == True:
-            #send zero before finishing
-            self.publishingFF = False
-            msg = Twist()
-            msg.linear.x = 0.0
-            msg.linear.y = 0.0
-            msg.linear.z = 0.0
-            msg.angular.x = 0.0
-            msg.angular.y = 0.0
-            msg.angular.z = 0.0
-
-            self.ffPub.publish(msg)
-
-    def ffAutoTuneCB(self, msg):
-        #write the autotune save if it has changed
-
-        #if writing is disabled
-        if not self.writeFFAutoTune:
-            return
         
-        #if the overseer has not initialized the controller yet
-        if not self.model_nodes["complete_controller"].loaded_params:
-            return
+        # # publish ff if necessary
+        # if not (self.get_parameter(FF_PUBLISH_PARAM).value):
+        #     #publish the ff
+        #     self.publishingFF = True
+
+        #     msg = Twist()
+        #     msg.linear.x = self.base_wrench[0]
+        #     msg.linear.y = self.base_wrench[1]
+        #     msg.linear.z = self.base_wrench[2]
+        #     msg.angular.x = self.base_wrench[3]
+        #     msg.angular.y = self.base_wrench[4]
+        #     msg.angular.z = self.base_wrench[5]
+
+        #     self.ffPub.publish(msg)
+        # elif self.publishingFF == True:
+        #     #send zero before finishing
+        #     self.publishingFF = False
+        #     msg = Twist()
+        #     msg.linear.x = 0.0
+        #     msg.linear.y = 0.0
+        #     msg.linear.z = 0.0
+        #     msg.angular.x = 0.0
+        #     msg.angular.y = 0.0
+        #     msg.angular.z = 0.0
+
+        #     self.ffPub.publish(msg)
+
+    # def ffAutoTuneCB(self, msg):
+    #     #write the autotune save if it has changed
+
+    #     #if writing is disabled
+    #     if not self.writeFFAutoTune:
+    #         return
+        
+    #     #if the overseer has not initialized the controller yet
+    #     if not self.model_nodes["complete_controller"].loaded_params:
+    #         return
     
-        #if the auto ff config path doesn't exist
-        if (self.autoff_config_path == None):
-            return 
+    #     #if the auto ff config path doesn't exist
+    #     if (self.autoff_config_path == None):
+    #         return 
 
-        #if the controller is trying to reinit the autoff
-        if(self.waiting_on_init):
+    #     #if the controller is trying to reinit the autoff
+    #     if(self.waiting_on_init):
 
-            auto_ff_init = None
-            try:
-                auto_ff_init = self.configTree["controller"]["autoff"]["initial_ff"]
-            except KeyError:
-                self.get_logger().warn("Cannot find the initial auto ff in config tree!")
-                return
+    #         auto_ff_init = None
+    #         try:
+    #             auto_ff_init = self.configTree["controller"]["autoff"]["initial_ff"]
+    #         except KeyError:
+    #             self.get_logger().warn("Cannot find the initial auto ff in config tree!")
+    #             return
 
-            if (abs(auto_ff_init[0] - msg.linear.x) < AUTOFF_INIT_TOLERANCE and abs(auto_ff_init[1] - msg.linear.y) < AUTOFF_INIT_TOLERANCE and abs(auto_ff_init[2] - msg.linear.z) < AUTOFF_INIT_TOLERANCE and 
-                abs(auto_ff_init[3] - msg.angular.x) < AUTOFF_INIT_TOLERANCE and abs(auto_ff_init[4] - msg.angular.y) < AUTOFF_INIT_TOLERANCE and abs(auto_ff_init[5] - msg.angular.z) < AUTOFF_INIT_TOLERANCE):
+    #         if (abs(auto_ff_init[0] - msg.linear.x) < AUTOFF_INIT_TOLERANCE and abs(auto_ff_init[1] - msg.linear.y) < AUTOFF_INIT_TOLERANCE and abs(auto_ff_init[2] - msg.linear.z) < AUTOFF_INIT_TOLERANCE and 
+    #             abs(auto_ff_init[3] - msg.angular.x) < AUTOFF_INIT_TOLERANCE and abs(auto_ff_init[4] - msg.angular.y) < AUTOFF_INIT_TOLERANCE and abs(auto_ff_init[5] - msg.angular.z) < AUTOFF_INIT_TOLERANCE):
 
-                #if the init signal has taken
-                self.waiting_on_init = False
+    #             #if the init signal has taken
+    #             self.waiting_on_init = False
 
-            else: 
-                #pub msg
-                msg = Empty()
-                self.re_init_signal_pub.publish(msg)
+    #         else: 
+    #             #pub msg
+    #             msg = Empty()
+    #             self.re_init_signal_pub.publish(msg)
 
-            return
+    #         return
 
-        #if the twist has been updated
-        if (not (self.currentAutoTuneTwist[0] == msg.linear.x and self.currentAutoTuneTwist[1] == msg.linear.y and self.currentAutoTuneTwist[2] == msg.linear.z and 
-           self.currentAutoTuneTwist[3] == msg.angular.x and self.currentAutoTuneTwist[4] == msg.angular.y and self.currentAutoTuneTwist[5] == msg.angular.z)) or (self.refresh_drag_file):
-            #prepare string to be wrote
-            auto_ff_config_string = f"auto_ff: [{msg.linear.x},{msg.linear.y},{msg.linear.z},{msg.angular.x},{msg.angular.y},{msg.angular.z}]\n"
+    #     #if the twist has been updated
+    #     if (not (self.currentAutoTuneTwist[0] == msg.linear.x and self.currentAutoTuneTwist[1] == msg.linear.y and self.currentAutoTuneTwist[2] == msg.linear.z and 
+    #        self.currentAutoTuneTwist[3] == msg.angular.x and self.currentAutoTuneTwist[4] == msg.angular.y and self.currentAutoTuneTwist[5] == msg.angular.z)) or (self.refresh_drag_file):
+    #         #prepare string to be wrote
+    #         auto_ff_config_string = f"auto_ff: [{msg.linear.x},{msg.linear.y},{msg.linear.z},{msg.angular.x},{msg.angular.y},{msg.angular.z}]\n"
             
-            drag_forward_string = ""
-            drag_reverse_string = ""
-            if(not ((self.drag_comp_forward_data is None) or (self.drag_comp_reverse_data is None))):
+    #         drag_forward_string = ""
+    #         drag_reverse_string = ""
+    #         if(not ((self.drag_comp_forward_data is None) or (self.drag_comp_reverse_data is None))):
 
-                #write the forward drag string
-                drag_forward_string = f"drag_forward: ["
-                for value in self.drag_comp_forward_data:
-                    drag_forward_string = drag_forward_string + str(value) + ","
-                drag_forward_string =  drag_forward_string + "]\n"
+    #             #write the forward drag string
+    #             drag_forward_string = f"drag_forward: ["
+    #             for value in self.drag_comp_forward_data:
+    #                 drag_forward_string = drag_forward_string + str(value) + ","
+    #             drag_forward_string =  drag_forward_string + "]\n"
 
-                #write the reverse drag string
-                drag_reverse_string = f"drag_reverse: ["
-                for value in self.drag_comp_reverse_data:
-                    drag_reverse_string = drag_reverse_string + str(value) + ","
-                drag_reverse_string =  drag_reverse_string + "]\n"
+    #             #write the reverse drag string
+    #             drag_reverse_string = f"drag_reverse: ["
+    #             for value in self.drag_comp_reverse_data:
+    #                 drag_reverse_string = drag_reverse_string + str(value) + ","
+    #             drag_reverse_string =  drag_reverse_string + "]\n"
 
-            #write config to file
-            try:
-                with open(self.autoff_config_path, "w") as config:
+    #         #write config to file
+    #         try:
+    #             with open(self.autoff_config_path, "w") as config:
                     
-                    config.write(auto_ff_config_string)
+    #                 config.write(auto_ff_config_string)
                     
 
-                    if not ((self.drag_comp_forward_data is None) or (self.drag_comp_reverse_data is None)):
-                     #cant save until cb complete
-                        config.write(drag_forward_string)
-                        config.write(drag_reverse_string)
+    #                 if not ((self.drag_comp_forward_data is None) or (self.drag_comp_reverse_data is None)):
+    #                  #cant save until cb complete
+    #                     config.write(drag_forward_string)
+    #                     config.write(drag_reverse_string)
                         
-                    config.close()
+    #                 config.close()
 
-            except FileExistsError:
-                self.get_logger().error(f"Cannot open ff auto tune file at: {self.autoff_config_path}")
-            except PermissionError:
-                self.get_logger().error(f"No Permission to write of autoff file at: {self.autoff_config_path}")
+    #         except FileExistsError:
+    #             self.get_logger().error(f"Cannot open ff auto tune file at: {self.autoff_config_path}")
+    #         except PermissionError:
+    #             self.get_logger().error(f"No Permission to write of autoff file at: {self.autoff_config_path}")
